4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -9,6 +9,11 @@ blogRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
+blogRouter.get('/:id', async (request, response) => {
+    const blogs = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1})
+    response.json(blogs)
+})
+
 blogRouter.post('/', async (request, response) => {
     const body = new Blog(request.body)
 
@@ -38,6 +43,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() !== decodedToken.id.toString()){
+        return response.status(403).json({
+            error: 'acces denied'
+        })
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
